+++ fastai/text/data.py
@@ -14,8 +14,8 @@ text_extensions = {'.txt'}
 class LanguageModelLoader():
     "Create a dataloader with bptt slightly changing."
     def __init__(self, dataset:LabelList, bs:int=64, bptt:int=70, backwards:bool=False, shuffle:bool=False,
-                 max_len:int=25):
-        self.dataset,self.bs,self.bptt,self.backwards,self.shuffle = dataset,bs,bptt,backwards,shuffle
+                 max_len:int=25, p_bptt:int=0.95):
+        self.dataset,self.bs,self.bptt,self.backwards,self.shuffle,self.p_bptt = dataset,bs,bptt,backwards,shuffle,p_bptt
         self.first,self.i,self.iter = True,0,0
         self.n = len(np.concatenate(dataset.x.items)) // self.bs if len(dataset.x.items) > 0 else 0
         self.max_len,self.num_workers = max_len,0
@@ -30,7 +30,7 @@ class LanguageModelLoader():
         while self.i < self.n-1 and self.iter<len(self):
             if self.first and self.i == 0: self.first,seq_len = False,self.bptt + self.max_len
             else:
-                bptt = self.bptt if np.random.random() < 0.95 else self.bptt / 2.
+                bptt = self.bptt if np.random.random() < self.p_bptt else self.bptt / 2.
                 seq_len = max(5, int(np.random.normal(bptt, 5)))
                 seq_len = min(seq_len, self.bptt + self.max_len)
             res = self.get_batch(self.i, seq_len)
@@ -38,7 +38,7 @@ class LanguageModelLoader():
             self.iter += 1
             yield res
 
-    def __len__(self) -> int: return (self.n-1) // self.bptt
+    def __len__(self) -> int: return int(math.ceil((self.n-1) / self.bptt))
     def __getattr__(self,k:str)->Any: return getattr(self.dataset, k)
 
     @property
@@ -56,7 +56,7 @@ class LanguageModelLoader():
     def get_batch(self, i:int, seq_len:int) -> Tuple[LongTensor, LongTensor]:
         "Create a batch at `i` of a given `seq_len`."
         seq_len = min(seq_len, len(self.data) - 1 - i)
-        return self.data[i:i+seq_len], self.data[i+1:i+1+seq_len].contiguous().view(-1)
+        return self.data[i:i+seq_len], self.data[i+1:i+1+seq_len]#.contiguous().view(-1)
 
 class SortSampler(Sampler):
     "Go through the text data by order of length."